move blobs-saver command as an ethrex l2 subcommand
