+++ cmd/ethrex/Cargo.toml
@@ -33,6 +33,8 @@ tokio-util.workspace = true
 redb = { workspace = true, optional = true }
 lazy_static.workspace = true
 secp256k1 = { workspace = true, optional = true }
+keccak-hash.workspace = true
+reqwest.workspace = true
 
 cfg-if = "1.0.0"
 
@@ -51,21 +53,21 @@ path = "./lib.rs"
 default = ["libmdbx", "c-kzg", "blst"]
 dev = ["dep:ethrex-dev"]
 c-kzg = [
-    "ethrex-vm/c-kzg",
-    "ethrex-common/c-kzg",
-    "ethrex-blockchain/c-kzg",
-    "ethrex-p2p/c-kzg",
+  "ethrex-vm/c-kzg",
+  "ethrex-common/c-kzg",
+  "ethrex-blockchain/c-kzg",
+  "ethrex-p2p/c-kzg",
 ]
 metrics = ["ethrex-blockchain/metrics"]
 libmdbx = ["ethrex-storage/libmdbx"]
 redb = ["dep:redb", "ethrex-storage/redb"]
 blst = ["ethrex-vm/blst"]
 l2 = [
-    "dep:ethrex-l2",
-    "ethrex-vm/l2",
-    "ethrex-blockchain/l2",
-    "ethrex-rpc/l2",
-    "dep:secp256k1",
+  "dep:ethrex-l2",
+  "ethrex-vm/l2",
+  "ethrex-blockchain/l2",
+  "ethrex-rpc/l2",
+  "dep:secp256k1",
 ]
 based = ["l2", "ethrex-rpc/based"]
